Move secureDocument to conformanceTools.js.
