+++ conformanceTools.js
@@ -1,9 +1,4 @@
-/*
-    Steps to create a signed verifiable credential with a simple
-    proof chain using Ed25519 signatures.
-*/
-
-import { mkdir, writeFile } from 'fs/promises';
+import { writeFile } from 'fs/promises';
 import { createRequire } from 'module';
 import jsonld from 'jsonld';
 import { localLoader } from './documentLoader.js';
@@ -12,51 +7,35 @@ import {ed25519 as ed} from '@noble/curves/ed25519';
 import { sha256 } from '@noble/hashes/sha256';
 import { bytesToHex, concatBytes } from '@noble/hashes/utils';
 
-
-const require = createRequire(import.meta.url);
-
-// Create output directory for the results
-const baseDir = "./output/eddsa-rdfc-2022/conformance/";
-// recursively create the dirs for the baseDir is needed
-let status = await mkdir(baseDir, {recursive: true});
-
-jsonld.documentLoader = localLoader; // Local loader for JSON-LD
-
-const keyPairs = require('./input/multiKeyPairs.json');
-    
-
-// Read input documents from files.
-const documents = new Map([
-  ['1.1', require('./input/v1/unsecured.json')],
-  ['2.0', require('./input/v2/unsecured.json')]
-]);
-
-
-async function secureDocument({
+export async function secureDocument({
+  baseDir,
+  chainKeys,
   credential,
   proofIds = ["urn:uuid:26329423-bec9-4b2e-88cb-a7c7d9dc4544"],
   fileName,
-  previousProofType
+  previousProofType,
+  debug = false
 }) {
   const document = structuredClone(credential);
-  // Signed Document Creation Steps:
-  
-  // Canonize the document
-  let cannon = await jsonld.canonize(document);
-  // console.log("Canonized unsigned document:")
-  // console.log(cannon);
-  // writeFile(baseDir + 'canonDocDataInt.txt', cannon);
-  
-  // Hash canonized document
-  let docHash = sha256(cannon); // @noble/hash will convert string to bytes via UTF-8
-  // console.log("Hash of canonized document in hex:")
-  // console.log(bytesToHex(docHash));
-  // writeFile(baseDir + 'docHashDataInt.txt', bytesToHex(docHash));
+  if(debug) {
+    // Signed Document Creation Steps:
+    
+    // Canonize the document
+    let cannon = await jsonld.canonize(document);
+    console.log("Canonized unsigned document:")
+    console.log(cannon);
+    writeFile(baseDir + 'canonDocDataInt.txt', cannon);
+    
+    // Hash canonized document
+    let docHash = sha256(cannon); // @noble/hash will convert string to bytes via UTF-8
+    console.log("Hash of canonized document in hex:")
+    console.log(bytesToHex(docHash));
+    writeFile(baseDir + 'docHashDataInt.txt', bytesToHex(docHash));
+  }
   
   
   // **Proof Chains** starting from document
   let signedDocument = structuredClone(document);
-  const chainKeys = [keyPairs.keyPair1, keyPairs.keyPair2];
   // set the first entry to null to prevent the first proof
   // from having a previousProof set
   const previousProofs = [null, proofIds[0]]; // Simple proof chain
@@ -105,10 +84,10 @@ async function secureDocument({
     console.log(matchingProofs);
     // Canonize the "chained" document
     writeFile(baseDir + `${fileName}-SimpleTempDoc${i+1}.json`, JSON.stringify(document, null, 2));
-    cannon = await jsonld.canonize(document);
+    const cannon = await jsonld.canonize(document);
   
     // Hash canonized chained document
-    docHash = sha256(cannon); // @noble/hash will convert string to bytes via UTF-8
+    const docHash = sha256(cannon); // @noble/hash will convert string to bytes via UTF-8
   
     // canonize the proof config
     let proofCanon = await jsonld.canonize(proofConfigChain);
@@ -173,29 +152,3 @@ function getVM(key) {
     '#' + key.publicKeyMultibase
 }
 
-// create versioned VCs with previousProof as string
-for(const [version, credential] of documents) {
-  await secureDocument({
-    credential,
-    fileName: `${version}-previousProofString`,
-    previousProofType: 'string'
-  });
-}
-// create versioned VCs with previousProof as an Array
-for(const [version, credential] of documents) {
-  await secureDocument({
-    credential,
-    fileName: `${version}-previousProofArray`,
-    previousProofType: 'Array'
-  });
-}
-
-// create versioned VCs with previousProof as an Array
-for(const [version, credential] of documents) {
-  await secureDocument({
-    credential,
-    fileName: `${version}-previousProofNotStringFail`,
-    previousProofType: 'string',
-    proofIds: [456321]
-  });
-}
